Added support to "mirrors". [INCOMPLETE]
Package working copies will be relocated to mirrors by default, and
temporarly relocated to the main servers when performing write operations
(such as ci and editlog).

Note that this task is not completely finished because repsys-1.6.X uses
pexpect and it leaves unable to use VISUAL to edit commit messages. This
is REALLY bad. Alternatives are:

- call VISUAL before calling svn.commit(), 
- another whacky workaround related to svn + pexpect, and
- don't add such feature to a (threoretically) stable branch.

I choose the later. This is a really deep change for a branch that should
be touched only for really needed features. Using pysvn we can do it in an
easier way.

This code just lays as a reference.