(effectively) Hide authors with only silented changesets
Now authors and/or releases will be effectively be hidden when they have
revisions/authors without visible lines. Note the revision number of
visible releases can be affected by silented revisions (those that are
newer than the latest visible)

This commit is too atomic, it could be splited in at least four changesets.