Reverted "Added configuration option confirm-revision to submit section"
It was a really bad idea. Another configuration option to help things to
get messy. The usage we would be trying to protect is new anyway, so users
will have to get used to this behavior.

Instead, the next revision should just warn the user about the revision
being submitted and that's all.