Don't say "invalid command CMD" when this is not the real error
Sometimes when a module used by repsys is missing, the given error message
is "invalid command CMD", instead of a exception trace showing that there's
something wrong with with the repsys setup.

This change fixes this issue by counting the number of frames in the
exception trace. When the command module is not available there will be
only one frame (being the function that is trying to import the command's
module), otherwise there will be more frames (because the found module will
already be present in the trace).

This fix is probably valid for smart too.