First simple working version of the upload structure
It contains a "upload" command that will upload the "sources" file and
will call a wrapper script that will run the right rsync command to a
given location.

With this approach we can have some kind of separation between the generic
code and the upload stuff, which currently is not well defined.