Check the files *before* copying them to the releases directory
We can't do it after the copy because we don't know how to revert the
transation (which directories to remove, as we don't have any knoledge
about the package structure).

Then in order to do it in a more-nebulous-so-that-I-can't-see-any- problem
way, I create one temporary directory in the repository root (to allow
hardlinks) and then copy there, check them and them copy to the final
location.