Make 'sync' handle the binrepo scheme
It required to emule some kind of "pending to be added" behavior of svn, as
sync should not commit changes unless one --commit option is used (not
implemented yet). It was implemented by creating a list of pending files at
SOURCES/.svn/repsys-pending. This seems to be a bad choice and will be a
headache in the future, let's see.

There's also the initial version of a commit() version that uses the
pending list and should be called by the rpmutil.commit function soon.