Add configuration option to sort the generated changelog
The boolean option "sort" in the [log] section can be used to make repsys
sort all the entries after generating the changelog from svn and
concatenating with the one found in "misc/".

It was also added another option to merge the log found in the spec (those
that don't have the %changelog section stripped).

The current problem with this implementation is that SILENTed entries in
svn will not be merged with those found in %changelog or in misc. Not a
problem really.