Rewrote the code that deals with file copies in get_srpm
Removed all the weird, unused or repeated code. 

It also now copies all the SRPMS found to the target directories, in order
to not rely on a guess (it was copying the first .src.rpm found).

The dangerous change is that I removed the use of "cp -f" to copy the
files. I tried to ensure the code behaves as the "-f" tell, but I am not
confident it will be always the case.