Introduced RepSys.layout and repository_url
repository_url (which has a much better name) will take care of generating
an URL from a package name, a relative package name, or another URL.

Now all commands that accept package names also accept the form
DISTRO/NAME, which will map to a URL inside updates/.

It also introduced a new configuration option: "repository". It defines the
ROOT of the subversion repository. With this option we can build URLs on
updates/. If this option is not set, it will try to build the URL from
default_parent for compatibility.