Added support for multiple binrepos
A section "[binrepos]" in the configuration file will have entries mapping
paths in the main text repository with the URLs of the real binrepos that
should be used, for example:

repository = file:///repos/svn/

[binrepos]
updates/2009.0 = file:///repos/svn-20090/
updates/2009.1 = file:///repos/svn-20091/

A URL file:///repos/svn/updates/2009.1/trafshow/current would be translated
to file:///repos/svn-20091/updates/2009.1/trafshow/current.

This change required the centralization of the function responsible for
translating URLs from the main text repo to the binrepo. It allowed to
simplify the funcionality of binrepo_url and then giving a better name to
it. Blablabla.

This feature is cool, but raises the issue of users having to update their
configuration files frequently in order access the right repositories.

Oh, see! More than a twitter post!